refactor: extract PlaybookSourceLinkageMixin from PlaybookMixin (sqlite + ABC)
Move the 5 agent-playbook source-linkage methods
(set/get_source_user_playbook_ids_for_agent_playbook[s],
set/get_source_windows_for_agent_playbook) verbatim out of the shrinking
PlaybookMixin into a dedicated PlaybookSourceLinkageMixin in the playbook/
subpackage, composed before PlaybookMixin. Behavior-preserving; bodies are
byte-identical. Task 6 of the playbook-storage decomposition.
