+++ reflexio/server/services/storage/sqlite_storage/playbook/__init__.py
@@ -1,4 +1,9 @@
 from ._agent import AgentPlaybookStoreMixin
+from ._source_linkage import PlaybookSourceLinkageMixin
 from ._user import UserPlaybookStoreMixin
 
-__all__ = ["AgentPlaybookStoreMixin", "UserPlaybookStoreMixin"]
+__all__ = [
+    "AgentPlaybookStoreMixin",
+    "PlaybookSourceLinkageMixin",
+    "UserPlaybookStoreMixin",
+]
